+++ MatrixSDK/Crypto/Data/Store/MXRealmCryptoStore/MXRealmCryptoStore.m
@@ -25,7 +25,7 @@
 #import "MXTools.h"
 #import "MXCryptoTools.h"
 
-NSUInteger const kMXRealmCryptoStoreVersion = 14;
+NSUInteger const kMXRealmCryptoStoreVersion = 15;
 
 static NSString *const kMXRealmCryptoStoreFolder = @"MXRealmCryptoStore";
 
@@ -145,6 +145,19 @@ RLM_ARRAY_TYPE(MXRealmOlmInboundGroupSession)
 }
 @end
 RLM_ARRAY_TYPE(MXRealmOlmOutboundGroupSession)
+
+@interface MXRealmSharedOutboundSession : RLMObject
+
+@property NSString *roomId;
+@property NSString *sessionId;
+@property MXRealmDeviceInfo *device;
+@property NSNumber<RLMInt> *messageIndex;
+
+@end
+
+@implementation MXRealmSharedOutboundSession
+@end
+RLM_ARRAY_TYPE(MXRealmSharedOutboundSession)
 
 @interface MXRealmOlmAccount : RLMObject
 
@@ -1046,7 +1059,6 @@ RLM_ARRAY_TYPE(MXRealmSecret)
         storedSession = [[MXOlmOutboundGroupSession alloc] initWithSession:session roomId:roomId creationTime:realmSession.creationTime];
     }];
 
-
     NSLog(@"[MXRealmCryptoStore] storeOutboundGroupSession: store 1 key (%lu new) in %.3fms", newCount, [[NSDate date] timeIntervalSinceDate:startDate] * 1000);
     
     return storedSession;
@@ -1103,6 +1115,91 @@ RLM_ARRAY_TYPE(MXRealmSecret)
         [realm deleteObjects:realmSessions];
         NSLog(@"[MXRealmCryptoStore] removeOutboundGroupSessionWithRoomId%@: removed %lu entries", roomId, realmSessions.count);
     }];
+}
+
+- (void)storeSharedDevices:(MXUsersDevicesMap<NSNumber *> *)devices messageIndex:(NSUInteger) messageIndex forOutboundGroupSessionInRoomWithId:(NSString *)roomId sessionId:(NSString *)sessionId
+{
+    NSDate *startDate = [NSDate date];
+
+    RLMRealm *realm = self.realm;
+
+    [realm transactionWithBlock:^{
+        
+        for (NSString *userId in [devices userIds])
+        {
+            for (NSString *deviceId in [devices deviceIdsForUser:userId])
+            {
+                MXRealmUser *realmUser = [MXRealmUser objectsInRealm:realm where:@"userId = %@", userId].firstObject;
+                if (!realmUser)
+                {
+                    NSLog(@"[MXRealmCryptoStore] storeSharedDevices cannot find user with the ID %@", userId);
+                    continue;
+                }
+
+                MXRealmDeviceInfo *realmDevice = [[realmUser.devices objectsWhere:@"deviceId = %@", deviceId] firstObject];
+                if (!realmDevice)
+                {
+                    NSLog(@"[MXRealmCryptoStore] storeSharedDevices cannot find device with the ID %@", deviceId);
+                    continue;
+                }
+
+                MXRealmSharedOutboundSession *sharedInfo = [[MXRealmSharedOutboundSession alloc] initWithValue: @{
+                    @"roomId": roomId,
+                    @"sessionId": sessionId,
+                    @"device": realmDevice,
+                    @"messageIndex": @(messageIndex)
+                }];
+                [realm addObject:sharedInfo];
+            }
+        }
+    }];
+
+    NSLog(@"[MXRealmCryptoStore] storeSharedDevices (count: %tu) in %.3fms", devices.count, [[NSDate date] timeIntervalSinceDate:startDate] * 1000);
+}
+
+- (MXUsersDevicesMap<NSNumber *> *)sharedDevicesForOutboundGroupSessionInRoomWithId:(NSString *)roomId sessionId:(NSString *)sessionId
+{
+    NSDate *startDate = [NSDate date];
+
+    MXUsersDevicesMap<NSNumber *> *devices = [MXUsersDevicesMap new];
+
+    RLMRealm *realm = self.realm;
+    
+    RLMResults<MXRealmSharedOutboundSession *> *results = [MXRealmSharedOutboundSession objectsInRealm:realm where:@"roomId = %@ AND sessionId = %@", roomId, sessionId];
+    
+    for (MXRealmSharedOutboundSession *sharedInfo in results)
+    {
+        MXDeviceInfo *deviceInfo = [NSKeyedUnarchiver unarchiveObjectWithData:sharedInfo.device.deviceInfoData];
+        if (!deviceInfo)
+        {
+            NSLog(@"[MXRealmCryptoStore] sharedDevicesForOutboundGroupSessionInRoomWithId cannot unarchive deviceInfo");
+            continue;
+        }
+        [devices setObject:sharedInfo.messageIndex forUser:deviceInfo.userId andDevice:deviceInfo.deviceId];
+    }
+
+    NSLog(@"[MXRealmCryptoStore] sharedDevicesForOutboundGroupSessionInRoomWithId (count: %tu) in %.3fms", results.count, [[NSDate date] timeIntervalSinceDate:startDate] * 1000);
+
+    return devices;
+}
+
+- (NSNumber *)messageIndexForSharedDeviceInRoomWithId:(NSString *)roomId sessionId:(NSString *)sessionId userId:(NSString *)userId deviceId:(NSString *)deviceId
+{
+    NSNumber *messageIndex = nil;
+    RLMRealm *realm = self.realm;
+    
+    RLMResults<MXRealmSharedOutboundSession *> *sessions = [MXRealmSharedOutboundSession objectsInRealm:realm where:@"roomId = %@ AND sessionId = %@ AND device.deviceId = %@", roomId, sessionId, deviceId];
+    for (MXRealmSharedOutboundSession *session in sessions)
+    {
+        MXDeviceInfo *deviceInfo = [NSKeyedUnarchiver unarchiveObjectWithData:session.device.deviceInfoData];
+        if ([deviceInfo.userId isEqualToString:userId])
+        {
+            messageIndex = session.messageIndex;
+            break;
+        }
+    }
+
+    return messageIndex;
 }
 
 #pragma mark - Key backup
@@ -1445,7 +1542,8 @@ RLM_ARRAY_TYPE(MXRealmSecret)
                              MXRealmOutgoingRoomKeyRequest.class,
                              MXRealmIncomingRoomKeyRequest.class,
                              MXRealmSecret.class,
-                             MXRealmOlmOutboundGroupSession.class
+                             MXRealmOlmOutboundGroupSession.class,
+                             MXRealmSharedOutboundSession.class
                              ];
 
     config.schemaVersion = kMXRealmCryptoStoreVersion;
@@ -1657,6 +1755,11 @@ RLM_ARRAY_TYPE(MXRealmSecret)
                 case 13:
                 {
                     NSLog(@"[MXRealmCryptoStore] Migration from schema #13 -> #14: Nothing to do (added MXRealmOlmOutboundGroupSession)");
+                }
+                    
+                case 14:
+                {
+                    NSLog(@"[MXRealmCryptoStore] Migration from schema #14 -> #15: Nothing to do (added MXRealmSharedOutboundSession)");
                 }
             }
         }